% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_plot.R
\name{flush_plot}
\alias{flush_plot}
\title{flush_plot}
\usage{
flush_plot(x, title, subtitle, caption)
}
\arguments{
\item{x}{ggplot object}

\item{title}{Grafiktitel}

\item{subtitle}{Untertitel}

\item{caption}{Quellenangabe}
}
\description{
plot function to left align title, subtitle and caption of ggplot object
}
\examples{
flush_plot(ggplot_object, title = "Titel", subtitle = "Untertitel", caption = "Datenquelle")

}
\keyword{flush_plot,}
\keyword{plot}
