% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swr.r
\name{swr}
\alias{swr}
\title{swr}
\usage{
swr(string, nwrap = 20)
}
\arguments{
\item{string}{string to break up}

\item{nwrap}{max character length per line}
}
\description{
string wrapper: break label, character string in multiple lines
}
\examples{
swr("Guten Morgen Zürich", 5)

}
\keyword{string,}
\keyword{wrap}
